refactor: 홈 상세 뷰와 디펜던시를 분리
