+++ Projects/Features/HomeFeature/Sources/Presentation/HomeFeatureView.swift
@@ -350,7 +350,7 @@ private extension HomeFeatureView {
     }
 }
 
-private struct HomeFeatureLoadingOverlay: View {
+struct HomeFeatureLoadingOverlay: View {
     var body: some View {
         ZStack {
             Color.mainBlack.opacity(0.08)
@@ -662,7 +662,7 @@ private struct ListKitComingPopupCell: View {
     }
 }
 
-private struct ListKitGridPopupCell: View {
+struct ListKitGridPopupCell: View {
     let popup: Popup
     let isLiked: Bool
     let cellWidth: CGFloat
@@ -787,106 +787,6 @@ private extension HomeFeatureView {
         if let targetPopup = allPopups.first(where: { $0.popupUuid == popupId }) {
             onSelectPopup(store.userUuid, targetPopup)
         }
-    }
-}
-
-public struct ComingPopupDetailFeatureView: View {
-    @State private var store: StoreOf<ComingPopupDetailReducer>
-    private let onSelectPopup: (String, Popup) -> Void
-
-    init(
-        store: StoreOf<ComingPopupDetailReducer>,
-        onSelectPopup: @escaping (String, Popup) -> Void = { _, _ in }
-    ) {
-        _store = State(initialValue: store)
-        self.onSelectPopup = onSelectPopup
-    }
-
-    public init(
-        userUuid: String,
-        popups: [Popup],
-        onSelectPopup: @escaping (String, Popup) -> Void = { _, _ in }
-    ) {
-        self.init(
-            store: Store(
-                initialState: ComingPopupDetailReducer.State(
-                    userUuid: userUuid,
-                    popups: popups
-                )
-            ) {
-                ComingPopupDetailReducer()
-            },
-            onSelectPopup: onSelectPopup
-        )
-    }
-
-    public var body: some View {
-        LKList {
-            LKSection(id: "coming-popup-detail") {
-                for popup in store.popups {
-                    LKRow(
-                        popup,
-                        id: \.popupUuid,
-                        reuseIdentifier: "HomeFeature.ListKitGridPopupCell"
-                    ) {
-                        ListKitGridPopupCell(
-                            popup: popup,
-                            isLiked: popup.isFavorited,
-                            cellWidth: HomeFeatureView.gridCellWidth,
-                            toggleLike: {
-                                store.send(.toggleLike(popup))
-                            }
-                        )
-                        .accessibilityElement(children: .ignore)
-                        .accessibilityIdentifier("home_comming_cell")
-                    }
-                    .equatableToken("\(popup.popupUuid)-\(popup.isFavorited)")
-                    .onSelect { _ in
-                        onSelectPopup(store.userUuid, popup)
-                    }
-                }
-            }
-            .sectionLayout(.grid(columns: 2, itemHeight: HomeFeatureView.gridCellHeight, columnSpacing: 15, rowSpacing: 20))
-            .sectionContentInsets(LKEdgeInsets(
-                top: 0,
-                left: .contentPadding,
-                bottom: 0,
-                right: .contentPadding
-            ))
-            .pinnedHeader(background: Color.subWhite)
-        }
-        .listKitStyle(.plain)
-        .updateEngine(.reloadData)
-        .scrollIndicators(.hidden)
-        .contentInsets(LKEdgeInsets(top: 0, left: 0, bottom: 0, right: 0))
-        .overlay {
-            if store.isLoading {
-                HomeFeatureLoadingOverlay()
-            }
-        }
-        .alert("안내", isPresented: isComingPopupErrorPresented) {
-            Button("확인", role: .cancel) {
-                store.send(.errorMessageChanged(nil))
-            }
-        } message: {
-            Text(store.errorMessage ?? "")
-        }
-        .onAppear {
-            store.send(.onAppear)
-        }
-    }
-}
-
-private extension ComingPopupDetailFeatureView {
-    var isComingPopupErrorPresented: Binding<Bool> {
-        Binding(
-            get: { store.errorMessage != nil },
-            set: { isPresented in
-                if !isPresented {
-                    store.send(.errorMessageChanged(nil))
-                }
-            }
-        )
     }
 }
